pyproject.toml
[tool.black]
skip-string-normalization = true

[tool.poetry]
name = "yearn-data-analytics"
version = "0.1.0"
description = ""
authors = [
    "borrito <yh01dlx@gmail.com>",
    "huggingbot <83656073+huggingbot@users.noreply.github.com>",
]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "0.20.0"
requests = "2.27.1"
web3 = "5.29.1"
pandas = "1.4.2"
jsons = "1.6.2"
sqlmodel = "0.0.6"
fastapi = "0.78.0"
uvicorn = "0.17.6"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
mypy = "0.950"
black = "22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
pre-commit-hooks = "^4.2.0"
types-requests = "^2.27.29"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
